v2
On the first level: texture fix
On the second level: added two rooms(eventually 3) where survivors can revive. Tanks now do not spawn after the elevator on the second map. Tanks or witch cannot spawn after the sliding bridge on the second map. Fixed the design of the elevator 2 map, now it does not fly in the air. 2 barrels now have collision. More arrows to point the way.

And I actually realized that testers should include regular people who haven't played L4D2 for 5000+ hours. Oh well, this is a change that will affect the entire design and balance of my future campaigns and addons.

+fix rescue trigger. now its disabled on spawn.